[SCB-1624]change starter name according to spring boot conventions.
