+++ archetypes/business-service-spring-boot2-starter/src/main/resources/archetype-resources/pom.xml
@@ -45,7 +45,7 @@
   <dependencies>
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>spring-boot2-starter-standalone</artifactId>
+      <artifactId>java-chassis-spring-boot-starter-standalone</artifactId>
     </dependency>
   </dependencies>
