+++ demo/demo-spring-boot-transport/demo-spring-boot-pojo-server/pom.xml
@@ -50,7 +50,7 @@
     </dependency>
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>spring-boot2-starter-servlet</artifactId>
+      <artifactId>java-chassis-spring-boot-starter-servlet</artifactId>
     </dependency>
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
