+++ integration-tests/it-producer-deploy-springboot2-servlet/pom.xml
@@ -45,7 +45,7 @@
     </dependency>
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>spring-boot2-starter-servlet</artifactId>
+      <artifactId>java-chassis-spring-boot-starter-servlet</artifactId>
     </dependency>
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
