+++ integration-tests/spring-jaxrs-tests/pom.xml
@@ -38,7 +38,7 @@
     </dependency>
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>spring-boot2-starter-servlet</artifactId>
+      <artifactId>java-chassis-spring-boot-starter-servlet</artifactId>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
