+++ java-chassis-distribution/pom.xml
@@ -132,29 +132,28 @@
     <!-- spring boot -->
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>spring-boot-common</artifactId>
+      <artifactId>java-chassis-spring-boot-common</artifactId>
     </dependency>
 
-    <!-- spring boot 2 -->
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>spring-boot2-starter</artifactId>
+      <artifactId>java-chassis-spring-boot-starter</artifactId>
     </dependency>
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>spring-boot2-starter-discovery</artifactId>
+      <artifactId>java-chassis-spring-boot-starter-discovery</artifactId>
     </dependency>
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>spring-boot2-starter-gateway</artifactId>
+      <artifactId>java-chassis-spring-boot-starter-gateway</artifactId>
     </dependency>
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>spring-boot2-starter-servlet</artifactId>
+      <artifactId>java-chassis-spring-boot-starter-servlet</artifactId>
     </dependency>
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>spring-boot2-starter-standalone</artifactId>
+      <artifactId>java-chassis-spring-boot-starter-standalone</artifactId>
     </dependency>
 
     <!-- metrics -->
