+++ java-chassis-spring-boot/java-chassis-spring-boot-common/pom.xml
@@ -20,24 +20,22 @@
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
-    <artifactId>java-chassis-parent</artifactId>
+    <artifactId>java-chassis-spring-boot</artifactId>
     <groupId>org.apache.servicecomb</groupId>
     <version>2.0.0-SNAPSHOT</version>
-    <relativePath>../../parents/default</relativePath>
   </parent>
+
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>spring-boot2-starter-parent</artifactId>
-  <name>Java Chassis::Spring Boot2 Starter::Parent</name>
-  <packaging>pom</packaging>
+  <artifactId>java-chassis-spring-boot-common</artifactId>
+  <name>Java Chassis::Foundations::Spring Boot Common</name>
 
-  <modules>
-    <module>spring-boot2-starter</module>
-    <module>spring-boot2-starter-servlet</module>
-    <module>spring-boot2-starter-standalone</module>
-    <module>spring-boot2-starter-discovery</module>
-    <module>spring-boot2-starter-gateway</module>   
-  </modules>
+  <dependencies>
+    <dependency>
+      <groupId>org.apache.servicecomb</groupId>
+      <artifactId>service-registry</artifactId>
+    </dependency>
+  </dependencies>
 
   <build>
     <plugins>
