+++ java-chassis-spring-boot/java-chassis-spring-boot-starters/java-chassis-spring-boot-starter-discovery/pom.xml
@@ -19,12 +19,13 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
+    <artifactId>java-chassis-spring-boot-starters</artifactId>
     <groupId>org.apache.servicecomb</groupId>
-    <artifactId>spring-boot2-starter-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>spring-boot2-starter-discovery</artifactId>
-  <name>Java Chassis::Spring Boot2 Starter::Discovery</name>
+
+  <artifactId>java-chassis-spring-boot-starter-discovery</artifactId>
+  <name>Java Chassis::Spring Boot Starter::Discovery</name>
 
   <dependencies>
     <dependency>
@@ -88,11 +89,11 @@
     </dependency>
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>spring-boot2-starter</artifactId>
+      <artifactId>java-chassis-spring-boot-starter</artifactId>
     </dependency>
     <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>spring-boot-common</artifactId>
+      <artifactId>java-chassis-spring-boot-common</artifactId>
     </dependency>
   </dependencies>
 </project>
