+++ java-chassis-spring-boot/java-chassis-spring-boot-starters/java-chassis-spring-boot-starter-discovery/src/main/resources/META-INF/spring.factories
