+++ java-chassis-spring-boot/java-chassis-spring-boot-starters/java-chassis-spring-boot-starter-gateway/pom.xml
@@ -19,35 +19,44 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <artifactId>java-chassis-parent</artifactId>
-    <groupId>org.apache.servicecomb</groupId>
-    <version>2.0.0-SNAPSHOT</version>
-    <relativePath>../../parents/default</relativePath>
-  </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>spring-boot-common</artifactId>
-  <name>Java Chassis::Foundations::Spring Boot Common</name>
+  <parent>
+    <artifactId>java-chassis-spring-boot-starters</artifactId>
+    <groupId>org.apache.servicecomb</groupId>
+    <version>2.0.0-SNAPSHOT</version>
+  </parent>
+
+  <artifactId>java-chassis-spring-boot-starter-gateway</artifactId>
+  <name>Java Chassis::Spring Boot Starter::Gateway</name>
 
   <dependencies>
     <dependency>
+      <groupId>org.springframework.cloud</groupId>
+      <artifactId>spring-cloud-starter-gateway</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework.cloud</groupId>
+      <artifactId>spring-cloud-gateway-webflux</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>ch.qos.logback</groupId>
+      <artifactId>logback-classic</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>ch.qos.logback</groupId>
+      <artifactId>logback-core</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.apache.servicecomb</groupId>
-      <artifactId>service-registry</artifactId>
+      <artifactId>java-chassis-spring-boot-starter-discovery</artifactId>
+      <exclusions>
+        <exclusion>
+          <groupId>org.apache.servicecomb</groupId>
+          <artifactId>provider-springmvc</artifactId>
+        </exclusion>
+      </exclusions>
     </dependency>
   </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-
 </project>
