+++ java-chassis-spring-boot/java-chassis-spring-boot-starters/pom.xml
@@ -20,38 +20,36 @@
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
-    <artifactId>springmvc-tests</artifactId>
-    <groupId>org.apache.servicecomb.tests</groupId>
+    <artifactId>java-chassis-spring-boot</artifactId>
+    <groupId>org.apache.servicecomb</groupId>
     <version>2.0.0-SNAPSHOT</version>
   </parent>
+
   <modelVersion>4.0.0</modelVersion>
 
-  <name>Java Chassis::Integration Tests::Spring MVC General With Spring Boot</name>
-  <artifactId>springmvc-tests-general-with-springboot</artifactId>
+  <artifactId>java-chassis-spring-boot-starters</artifactId>
+  <name>Java Chassis::Spring Boot Starter::Parent</name>
+  <packaging>pom</packaging>
 
-  <dependencies>
-    <dependency>
-      <groupId>org.apache.servicecomb.tests</groupId>
-      <artifactId>springmvc-tests-common</artifactId>
-      <version>${project.version}</version>
-      <type>test-jar</type>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.servicecomb</groupId>
-      <artifactId>java-chassis-spring-boot-starter-servlet</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-test</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework</groupId>
-      <artifactId>spring-test</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.hibernate.validator</groupId>
-      <artifactId>hibernate-validator</artifactId>
-    </dependency>
-  </dependencies>
+  <modules>
+    <module>java-chassis-spring-boot-starter</module>
+    <module>java-chassis-spring-boot-starter-servlet</module>
+    <module>java-chassis-spring-boot-starter-standalone</module>
+    <module>java-chassis-spring-boot-starter-discovery</module>
+    <module>java-chassis-spring-boot-starter-gateway</module>   
+  </modules>
+
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <configuration>
+          <source>1.8</source>
+          <target>1.8</target>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
 
 </project>
