+++ java-chassis-spring-boot/pom.xml
@@ -20,18 +20,20 @@
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
-    <artifactId>java-chassis</artifactId>
     <groupId>org.apache.servicecomb</groupId>
+    <artifactId>java-chassis-parent</artifactId>
     <version>2.0.0-SNAPSHOT</version>
+    <relativePath>../parents/default</relativePath>
   </parent>
+
   <modelVersion>4.0.0</modelVersion>
   <packaging>pom</packaging>
   <artifactId>java-chassis-spring-boot</artifactId>
   <name>Java Chassis::Spring Boot</name>
 
   <modules>
-    <module>spring-boot-common</module>
-    <module>spring-boot2-starter-parent</module>
+    <module>java-chassis-spring-boot-common</module>
+    <module>java-chassis-spring-boot-starters</module>
   </modules>
 
 </project>
